chore(logging): :hammer: make sure it works

+++ packages/backend/convex/subscriptions.ts
@@ -1,16 +1,11 @@
-import { getAuthUserId } from "@convex-dev/auth/server";
-import { Polar } from "@polar-sh/sdk";
-import { v } from "convex/values";
-import { api, internal } from "./_generated/api";
-import {
-  action,
-  internalMutation,
-  internalQuery,
-  mutation,
-  query,
-} from "./_generated/server";
-import { env } from "./env";
-import schema from "./schema";
+import { getAuthUserId } from '@convex-dev/auth/server';
+import { Polar } from '@polar-sh/sdk';
+import { v } from 'convex/values';
+
+import { api, internal } from './_generated/api';
+import { action, internalMutation, internalQuery, mutation, query } from './_generated/server';
+import { env } from './env';
+import schema from './schema';
 
 const createCheckout = async ({
   customerEmail,
@@ -23,27 +18,23 @@ const createCheckout = async ({
   successUrl: string;
   subscriptionId?: string;
 }) => {
-  const polar = new Polar({
-    server: "sandbox",
-    accessToken: env.POLAR_ACCESS_TOKEN,
-  });
+  const polar = new Polar({ server: 'sandbox', accessToken: env.POLAR_ACCESS_TOKEN });
   const result = await polar.checkouts.create({
     productPriceId,
     successUrl,
     customerEmail,
     subscriptionId,
   });
+
   return result;
 };
 
 export const getPlanByKey = internalQuery({
-  args: {
-    key: schema.tables.plans.validator.fields.key,
-  },
+  args: { key: schema.tables.plans.validator.fields.key },
   handler: async (ctx, args) => {
     return ctx.db
-      .query("plans")
-      .withIndex("key", (q) => q.eq("key", args.key))
+      .query('plans')
+      .withIndex('key', (q) => q.eq('key', args.key))
       .unique();
   },
 });
@@ -51,56 +42,44 @@ export const getPlanByKey = internalQuery({
 export const getOnboardingCheckoutUrl = action({
   handler: async (ctx) => {
     const user = await ctx.runQuery(api.users.getUser);
-    if (!user) {
-      throw new Error("User not found");
-    }
-    const product = await ctx.runQuery(internal.subscriptions.getPlanByKey, {
-      key: "free",
-    });
+    if (!user) throw new Error('User not found');
+
+    const product = await ctx.runQuery(internal.subscriptions.getPlanByKey, { key: 'free' });
     const price = product?.prices.month?.usd;
-    if (!price) {
-      throw new Error("Price not found");
-    }
-    if (!user.email) {
-      throw new Error("User email not found");
-    }
+
+    if (!price) throw new Error('Price not found');
+    if (!user.email) throw new Error('User email not found');
+
     const checkout = await createCheckout({
       customerEmail: user.email,
       productPriceId: price.polarId,
       successUrl: `${env.SITE_URL}/settings/billing`,
     });
+
     return checkout.url;
   },
 });
 
 export const getProOnboardingCheckoutUrl = action({
-  args: {
-    interval: schema.tables.subscriptions.validator.fields.interval,
-  },
+  args: { interval: schema.tables.subscriptions.validator.fields.interval },
   handler: async (ctx, args) => {
     const user = await ctx.runQuery(api.users.getUser);
-    if (!user) {
-      throw new Error("User not found");
-    }
-    const product = await ctx.runQuery(internal.subscriptions.getPlanByKey, {
-      key: "pro",
-    });
+    if (!user) throw new Error('User not found');
+
+    const product = await ctx.runQuery(internal.subscriptions.getPlanByKey, { key: 'pro' });
     const price =
-      args.interval === "month"
-        ? product?.prices.month?.usd
-        : product?.prices.year?.usd;
-    if (!price) {
-      throw new Error("Price not found");
-    }
-    if (!user.email) {
-      throw new Error("User email not found");
-    }
+      args.interval === 'month' ? product?.prices.month?.usd : product?.prices.year?.usd;
+
+    if (!price) throw new Error('Price not found');
+    if (!user.email) throw new Error('User email not found');
+
     const checkout = await createCheckout({
       customerEmail: user.email,
       productPriceId: price.polarId,
       successUrl: `${env.SITE_URL}/settings/billing`,
       subscriptionId: user.subscription?.polarId,
     });
+
     return checkout.url;
   },
 });
@@ -108,60 +87,50 @@ export const getProOnboardingCheckoutUrl = action({
 export const listPlans = query({
   handler: async (ctx) => {
     const userId = await getAuthUserId(ctx);
-    if (!userId) {
-      throw new Error("User not found");
-    }
-    const plans = await ctx.db.query("plans").collect();
+    if (!userId) throw new Error('User not found');
+
+    const plans = await ctx.db.query('plans').collect();
+
     return plans.sort((a, b) => a.key.localeCompare(b.key));
   },
 });
 
 export const getPolarEventUser = internalQuery({
-  args: {
-    polarId: v.optional(v.string()),
-    email: v.string(),
-  },
+  args: { polarId: v.optional(v.string()), email: v.string() },
   handler: async (ctx, args) => {
     const userById = args.polarId
       ? await ctx.db
-          .query("users")
-          .withIndex("polarId", (q) => q.eq("polarId", args.polarId))
+          .query('users')
+          .withIndex('polarId', (q) => q.eq('polarId', args.polarId))
           .unique()
       : undefined;
     const user =
       userById ||
       (await ctx.db
-        .query("users")
-        .withIndex("email", (q) => q.eq("email", args.email))
+        .query('users')
+        .withIndex('email', (q) => q.eq('email', args.email))
         .unique());
-    if (!user) {
-      throw new Error("User not found");
-    }
-    if (user.polarId) {
-      throw new Error(`User ${user.email} already has a Polar ID`);
-    }
+
+    if (!user) throw new Error('User not found');
+    if (user.polarId) throw new Error(`User ${user.email} already has a Polar ID`);
+
     const subscription = await ctx.db
-      .query("subscriptions")
-      .withIndex("userId", (q) => q.eq("userId", user._id))
+      .query('subscriptions')
+      .withIndex('userId', (q) => q.eq('userId', user._id))
       .unique();
-    if (!subscription) {
-      return user;
-    }
+
+    if (!subscription) return user;
+
     const plan = await ctx.db.get(subscription.planId);
-    if (!plan) {
-      throw new Error("Plan not found");
-    }
-    return {
-      ...user,
-      subscription,
-      plan,
-    };
+    if (!plan) throw new Error('Plan not found');
+
+    return { ...user, subscription, plan };
   },
 });
 
 export const replaceSubscription = internalMutation({
   args: {
-    userId: v.id("users"),
+    userId: v.id('users'),
     subscriptionPolarId: v.string(),
     input: v.object({
       currency: schema.tables.subscriptions.validator.fields.currency,
@@ -176,22 +145,20 @@ export const replaceSubscription = internalMutation({
   },
   handler: async (ctx, args) => {
     const subscription = await ctx.db
-      .query("subscriptions")
-      .withIndex("userId", (q) => q.eq("userId", args.userId))
+      .query('subscriptions')
+      .withIndex('userId', (q) => q.eq('userId', args.userId))
       .unique();
-    if (subscription) {
-      await ctx.db.delete(subscription._id);
-    }
+
+    if (subscription) await ctx.db.delete(subscription._id);
+
     const plan = await ctx.db
-      .query("plans")
-      .withIndex("polarProductId", (q) =>
-        q.eq("polarProductId", args.input.productId),
-      )
+      .query('plans')
+      .withIndex('polarProductId', (q) => q.eq('polarProductId', args.input.productId))
       .unique();
-    if (!plan) {
-      throw new Error("Plan not found");
-    }
-    await ctx.db.insert("subscriptions", {
+
+    if (!plan) throw new Error('Plan not found');
+
+    await ctx.db.insert('subscriptions', {
       userId: args.userId,
       planId: plan._id,
       polarId: args.subscriptionPolarId,
@@ -203,41 +170,33 @@ export const replaceSubscription = internalMutation({
       currentPeriodEnd: args.input.currentPeriodEnd,
       cancelAtPeriodEnd: args.input.cancelAtPeriodEnd,
     });
+
     const user = await ctx.db.get(args.userId);
-    if (!user?.polarSubscriptionPendingId) {
-      return;
-    }
+    if (!user?.polarSubscriptionPendingId) return;
+
     await ctx.scheduler.cancel(user.polarSubscriptionPendingId);
-    await ctx.db.patch(args.userId, {
-      polarSubscriptionPendingId: undefined,
-    });
+    await ctx.db.patch(args.userId, { polarSubscriptionPendingId: undefined });
   },
 });
 
 export const setSubscriptionPending = mutation({
   handler: async (ctx, args) => {
     const userId = await getAuthUserId(ctx);
-    if (!userId) {
-      throw new Error("User not found");
-    }
+    if (!userId) throw new Error('User not found');
+
     const scheduledFunctionId = await ctx.scheduler.runAfter(
       1000 * 120,
       internal.subscriptions.unsetSubscriptionPending,
       { userId },
     );
-    await ctx.db.patch(userId, {
-      polarSubscriptionPendingId: scheduledFunctionId,
-    });
+
+    await ctx.db.patch(userId, { polarSubscriptionPendingId: scheduledFunctionId });
   },
 });
 
 export const unsetSubscriptionPending = internalMutation({
-  args: {
-    userId: v.id("users"),
-  },
+  args: { userId: v.id('users') },
   handler: async (ctx, args) => {
-    await ctx.db.patch(args.userId, {
-      polarSubscriptionPendingId: undefined,
-    });
+    await ctx.db.patch(args.userId, { polarSubscriptionPendingId: undefined });
   },
 });
